Fix running issue on docker container
