+++ backend/mock/index.js
@@ -1,126 +1,3 @@
-import raw from './raw';
+const catalogue = require('./catalogue');
 
-const dejiesa = {
-  id: 'a4488c17-1b8b-40d6-9b4e-ba0ac0fc71ab',
-  slug: 'dejiesa-plain-cold-shoulder-mini-dress',
-  descriptionMarkdown: 'Warna Blue dan Light Blue ( kiri kekanan )  \nSatu ukuran muat sampai size XL   \nLingkar dada 104 cm, Lingkar lengan 36 cm, Panjang 93 cm,  \nLingkar pinggang 102 cm, Lingkar pinggul 104 cm  \nBahan Jeans\n\nModel Menggunakan Size & Warna : XL',
-  title: 'Dejiesa Plain Cold Shoulder Mini Dress',
-  quantity: 13,
-  images: [
-    {
-      // id: 224196,
-      fullUrl: 'https://ss-imager-prod.freetls.fastly.net/www-images/480/product_images/cfb83b40b45029b484d400af378e55cd.jpg',
-      width: 800,
-      height: 800,
-    },
-  ],
-  price: {
-    amount: 159000,
-    currency: 'IDR',
-  },
-  category: 'Mini Dress',
-};
-
-const frixza = {
-  id: 'd9d2ba40-a1a9-4377-b3a0-cae50216e8f1',
-  title: 'Frixza Plain Flare Maxi Dress',
-  slug: 'frixza-plain-flare-maxi-dress',
-  descriptionMarkdown: 'Warna Black, Maroon, dan Navy  \nSatu ukuran muat sampai size S  \nLingkar dada 90 cm, Lingkar Pinggang 78 cm, Lingkar Pinggul 140 cm, Panjang 140 cm\n\n Satu ukuran muat sampai size M  \nLingkar dada 94 cm, Lingkar Pinggang 82 cm, Lingkar Pinggul 146 cm, Panjang 140 cm \n\n Satu ukuran muat sampai size L  \nLingkar dada 100 cm, Lingkar Pinggang 88 cm, Lingkar Pinggul 150 cm, Panjang 140 cm\n\n Satu ukuran muat sampai size XL  \nLingkar dada 110 cm, Lingkar Pinggang 98 cm, Lingkar Pinggul 154 cm, Panjang 140 cm   \n\nBahan Wool Crepe\n\nResleting belakang  \nModel menggunakan size M dan Navy',
-  quantity: 62,
-  price: {
-    amount: 169000,
-    currency: 'IDR',
-  },
-  category: 'Maxi Dress',
-  images: [
-    {
-      // id: 227752,
-      fullUrl: 'https://ss-imager-prod.freetls.fastly.net/www-images/480/product_images/caeffd8c014bf570f9e5633f96af0e48.jpg',
-      width: 800,
-      height: 800,
-    },
-  ],
-};
-
-const galzinsa = {
-  id: 'da9049b3-e7cf-4373-997e-d5c44bb09033',
-  title: 'Galzinsa Lace Bodycon Mini Dress',
-  slug: 'galzinsa-lace-bodycon-mini-dress',
-  descriptionMarkdown: 'Warna Black, Broken White, Green, Maroon, Navy, dan Pink (kiri ke kanan)  \nSize S :  \nLingkar dada 88 cm, Lebar bahu 35 cm, Panjang lengan 19 cm, Lingkar lengan 32 cm, Panjang 90 cm, Lingkar pinggang 78 cm, Lingkar pinggul 92 cm  \nSize M :  \nLingkar dada 92 cm, Lebar bahu 36 cm, Panjang lengan 20 cm, Lingkar lengan 34 cm, Panjang 90 cm, Lingkar pinggang 80 cm, Lingkar pinggul 96 cm  \nSize L :  \nLingkar dada 96 cm, Lebar bahu 37 cm, Panjang lengan 21 cm, Lingkar lengan 36 cm, Panjang 92 cm, Lingkar pinggang 84 cm, Lingkar pinggul 100 cm  \nSize XL :  \nLingkar dada 100 cm, Lebar bahu 38 cm, Panjang lengan 22 cm, Lingkar lengan 38 cm, Panjang 100 cm, Lingkar pinggang 88 cm, Lingkar pinggul 104 cm    \nBahan Brukat lapis Furing  \n1 Resleting belakang  \n\nModel Menggunakan Size M & Warna Navy',
-  quantity: 778,
-  price: {
-    amount: 139000,
-    currency: 'IDR',
-  },
-  category: 'Mini Dress',
-  images: [
-    {
-      // id: 227753,
-      fullUrl: 'https://ss-imager-prod.freetls.fastly.net/www-images/480/product_images/60d4bd5f59a6c89028f84678cef6f3b2.jpg',
-      width: 800,
-      height: 800,
-    },
-    {
-      // id: 227754,
-      fullUrl: 'https://ss-imager-prod.freetls.fastly.net/www-images/480/product_images/7ce7c0bd615a20033186de7d11475749.jpg',
-      width: 800,
-      height: 800,
-    },
-    {
-      // id: 227755,
-      fullUrl: 'https://ss-imager-prod.freetls.fastly.net/www-images/480/product_images/e0e21d8b6fb8071197babebc89995009.jpg',
-      width: 800,
-      height: 800,
-    },
-    {
-      // id: 227756,
-      fullUrl: 'https://ss-imager-prod.freetls.fastly.net/www-images/480/product_images/8ed0bd11f6c4878e889311e1a91cd0dc.jpg',
-      width: 800,
-      height: 800,
-    },
-    {
-      // id: 227757,
-      fullUrl: 'https://ss-imager-prod.freetls.fastly.net/www-images/480/product_images/652e63a21828ead238b34ee3bb52516d.jpg',
-      width: 800,
-      height: 800,
-    },
-    {
-      // id: 227758,
-      fullUrl: 'https://ss-imager-prod.freetls.fastly.net/www-images/480/product_images/38cfc91ee13817f546adcf87146626fb.jpg',
-      width: 800,
-      height: 800,
-    },
-    {
-      // id: 227759,
-      fullUrl: 'https://ss-imager-prod.freetls.fastly.net/www-images/480/product_images/fb14df572a946a5b97e8ec876d6c9f54.jpg',
-      width: 800,
-      height: 800,
-    },
-    {
-      // id: 227760,
-      fullUrl: 'https://ss-imager-prod.freetls.fastly.net/www-images/480/product_images/fa3b22cdd22efb1775ac0e606d634a14.jpg',
-      width: 800,
-      height: 800,
-    },
-  ],
-};
-
-const items = [
-  dejiesa,
-  frixza,
-  galzinsa,
-].concat(raw);
-
-const product = {
-  items,
-  total: items.length,
-};
-
-const catalogue = {
-  id: 'a2e458ad-28b3-11e5-8c52-061c7301af31',
-  title: 'Dress',
-  metaTitle: 'Baju Dress Gaun Wanita Model Panjang dan Pendek',
-  product,
-};
-
-export default catalogue;
+module.exports = { ...catalogue };
