+++ backend/package.json
@@ -10,7 +10,7 @@
     "debug": "node --nolazy --inspect-brk=9229 --require babel-register lib/index.js",
     "lint": "eslint lib mock resolvers schema",
     "lint:fix": "yarn lint -- --fix",
-    "build": "babel lib schema/** -d dist",
+    "build": "babel lib resolvers/** schema/** mock/** -d dist",
     "serve": "yarn && yarn build && node dist/index.js",
     "test": "mocha --require babel-register schema/**/*.spec.js"
   },
